Moved endpoints to administration and added api key validation
https://github.com/SoftUni-Internal/exam-systems-issues/issues/1559
