+++ Services/UI/OJS.Services.Ui.Business/IContestsBusinessService.cs
@@ -2,9 +2,6 @@ namespace OJS.Services.Ui.Business
 {
     using System.Collections.Generic;
     using System.Threading.Tasks;
-    using OJS.Data.Models.Contests;
-    using OJS.Services.Common.Models;
-    using OJS.Services.Common.Models.Contests;
     using OJS.Services.Ui.Models.Contests;
     using OJS.Services.Ui.Models.Search;
     using OJS.Services.Infrastructure.Models;
@@ -23,10 +20,6 @@ namespace OJS.Services.Ui.Business
         Task ValidateContestPassword(int id, bool official, string password);
 
         Task<ContestDetailsServiceModel> GetContestDetails(int id);
-
-        Task<ContestLegacyExportServiceModel> Export(int id);
-
-        Task<IEnumerable<int>> GetExistingIds(IEnumerable<int> ids);
 
         Task<ContestParticipationServiceModel> GetParticipationDetails(StartContestParticipationServiceModel model);
