+++ Services/UI/OJS.Services.Ui.Business/Implementations/ContestsBusinessService.cs
@@ -8,7 +8,6 @@ namespace OJS.Services.Ui.Business.Implementations
     using X.PagedList;
     using FluentExtensions.Extensions;
     using Microsoft.EntityFrameworkCore;
-    using Newtonsoft.Json;
     using OJS.Common;
     using OJS.Data.Models.Contests;
     using OJS.Data.Models.Participants;
@@ -23,7 +22,6 @@ namespace OJS.Services.Ui.Business.Implementations
     using OJS.Services.Ui.Data;
     using OJS.Services.Ui.Models.Contests;
     using OJS.Services.Ui.Models.Search;
-    using System.Diagnostics;
     using static OJS.Services.Common.Constants.PaginationConstants.Contests;
 
     public class ContestsBusinessService : IContestsBusinessService
@@ -144,72 +142,6 @@ namespace OJS.Services.Ui.Business.Implementations
             contestDetailsServiceModel.IsActive = await this.activityService.IsContestActive(activityServiceModel);
 
             return contestDetailsServiceModel;
-        }
-
-        public async Task<ContestLegacyExportServiceModel> Export(int id)
-        {
-            var contest = await this.contestsData
-                .GetByIdQuery(id)
-                .AsSplitQuery()
-                .Where(c => !c.IsDeleted)
-                .Include(c => c.ProblemGroups)
-                    .ThenInclude(pg => pg.Problems)
-                    .ThenInclude(p => p.Tests)
-                .Include(c => c.ProblemGroups)
-                    .ThenInclude(pg => pg.Problems)
-                    .ThenInclude(p => p.Checker)
-                .Include(c => c.ProblemGroups)
-                    .ThenInclude(pg => pg.Problems)
-                        .ThenInclude(p => p.SubmissionTypesInProblems)
-                            .ThenInclude(sp => sp.SubmissionType)
-                .Include(c => c.ProblemGroups)
-                    .ThenInclude(pg => pg.Problems)
-                    .ThenInclude(p => p.Resources)
-                .FirstOrDefaultAsync();
-
-            if (contest == null)
-            {
-                throw new BusinessServiceException($"Contest with Id:{id} not found.");
-            }
-
-            RemoveCircularReferences(contest);
-
-            return contest.Map<ContestLegacyExportServiceModel>();
-        }
-
-        public async Task<IEnumerable<int>> GetExistingIds(IEnumerable<int> ids)
-            => await this.contestsData
-                .GetQuery()
-                .Where(c => !c.IsDeleted && ids.Contains(c.Id))
-                .Select(c => c.Id)
-                .ToListAsync();
-
-        private static void RemoveCircularReferences(Contest contest)
-        {
-            foreach (var problemGroup in contest.ProblemGroups)
-            {
-                problemGroup.Contest = null;
-
-                foreach (var problem in problemGroup.Problems)
-                {
-                    problem.ProblemGroup = null;
-
-                    foreach (var test in problem.Tests)
-                    {
-                        test.Problem = null;
-                    }
-
-                    foreach (var resource in problem.Resources)
-                    {
-                        resource.Problem = null;
-                    }
-
-                    foreach (var submissionTypeInProblem in problem.SubmissionTypesInProblems)
-                    {
-                        submissionTypeInProblem.Problem = null;
-                    }
-                }
-            }
         }
 
         public async Task<ContestRegistrationDetailsServiceModel> GetContestRegistrationDetails(int id, bool isOfficial)
